refactor code, split business logic and routers by creating new services.py file with  in dependecies folder
